Homework 1 Submission
The aim of HW1 is to get started with Git and LaTeX and a bit of rewriting. Relevant resources are
https://codeberg.org/alexhkurz/programming-languages-2025/src/branch/main/syllabus-long.md   
https://codeberg.org/alexhkurz/programming-languages-2025/src/branch/main/assignments.md
https://codeberg.org/alexhkurz/programming-languages-2025/src/branch/main/git-best-practices.md
and, most importantly,
https://codeberg.org/alexhkurz/programming-languages-2025/src/branch/main/report.md  
https://codeberg.org/alexhkurz/programming-languages-2025/src/branch/main/lecture-by-lecture.md  
For HW1 write a section in the report about the MU puzzle and how to solve it. Create a private Github repository containing your `report.tex` and your `report.pdf`.
A good name for your repo would be CPSC-354.
Submit on Canvas the URL to the file report.pdf in your private Github repo. To invite me to your repo, I am jonweinb
Links to an external site.on Github.
Your repo must contain both report.tex and the pdf-file compiled from the tex-file. The deadline is Sunday so that we can go over your work on Monday and give feedback on Tuesday.
Ask an interesting question on Discord, in the channel dedicated to the section you are taking. Suitable topics include everything related to the lectures, the reading or the discrete mathematics background.